requirements.txt
azure-servicebus
azure-data-tables
beautifulsoup4
flask
flasgger
python-dotenv
